refactor Node storage remove graph_ops from GH benches
