+++ .github/workflows/benchmark.yml
@@ -44,7 +44,6 @@ jobs:
         run: |
           set -o pipefail
           cargo bench --bench base -p raphtory-benchmark -- --output-format=bencher | tee benchmark-result.txt
-          cargo bench --bench graph_ops -p raphtory-benchmark -- --output-format=bencher | tee benchmark-result.txt
           cargo bench --bench algobench -p raphtory-benchmark -- --output-format=bencher | tee benchmark-result.txt
       - name: Delete cargo.lock if it exists
         run: |
